+++ Screenbar/EventMonitor.swift
@@ -1,9 +1,6 @@
 import Cocoa
 
 open class EventMonitor {
-    // fileprivate means private in this file
-    // private means only used in class or structure
-    
     fileprivate var monitor: AnyObject?
     fileprivate let mask: NSEvent.EventTypeMask
     fileprivate let handler: (NSEvent?) -> ()
@@ -18,13 +15,11 @@ open class EventMonitor {
         stop()
     }
     
-    // open is kind of like public, but has greater useing than public
-    // ? means the type of this value can be null or others
     open func start() {
         monitor = NSEvent.addGlobalMonitorForEvents(matching: mask, handler: handler) as AnyObject?
     }
     
-    // ! maens the value is not null
+    
     open func stop() {
         if monitor != nil {
             NSEvent.removeMonitor(monitor!)